refactor/daily-log-card
