+++ components/TasksScreen/index.tsx
@@ -14,6 +14,7 @@ import Link from "next/link"
 import { Card, CardHeader, CardTitle, CardContent } from "@/components/ui/card"
 import { Button } from "@/components/ui/button"
 import { CurrentTaskCard } from "./CurrentTaskCard"
+import { DailyLogCard } from "./DailyLogCard";
 
 export function Tasks() {
 
@@ -51,6 +52,8 @@ export function Tasks() {
       </header>
       <main className="flex-1 p-8">
         <div className="max-w-4xl mx-auto grid grid-cols-1 md:grid-cols-2 gap-8">
+          
+          {/* Current Task Card */}
           <CurrentTaskCard
             icon="🚀"
             title="Develop new feature"
@@ -58,53 +61,11 @@ export function Tasks() {
             onStart={handleStart}
             onPause={handlePause}
           />
-          <Card>
-            <CardHeader>
-              <CardTitle>Daily Log</CardTitle>
-            </CardHeader>
-            <CardContent className="flex flex-col gap-4">
-              <div className="flex items-center justify-between">
-                <div className="flex items-center gap-2">
-                  <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">🕰️</div>
-                  <div>
-                    <h3 className="text-lg font-bold">Focused Work</h3>
-                    <p className="text-sm text-[#8b949e]">2 hours 15 minutes</p>
-                  </div>
-                </div>
-                <Button variant="secondary" size="sm">
-                  View
-                </Button>
-              </div>
-              <div className="flex items-center justify-between">
-                <div className="flex items-center gap-2">
-                  <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
-                    🍲
-                  </div>
-                  <div>
-                    <h3 className="text-lg font-bold">Lunch Break</h3>
-                    <p className="text-sm text-[#8b949e]">45 minutes</p>
-                  </div>
-                </div>
-                <Button variant="secondary" size="sm">
-                  View
-                </Button>
-              </div>
-              <div className="flex items-center justify-between">
-                <div className="flex items-center gap-2">
-                  <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
-                    💤
-                  </div>
-                  <div>
-                    <h3 className="text-lg font-bold">Break</h3>
-                    <p className="text-sm text-[#8b949e]">30 minutes</p>
-                  </div>
-                </div>
-                <Button variant="secondary" size="sm">
-                  View
-                </Button>
-              </div>
-            </CardContent>
-          </Card>
+
+          {/* Daily Log Card */}
+          <DailyLogCard />
+
+          {/* Productivity Card */}
           <Card>
             <CardHeader>
               <CardTitle>Productivity Stats</CardTitle>
